pyproject.toml
[project]
name = "git-ssh-gui"
version = "0.1.0"
description = "A GUI application for managing Git repositories on remote SSH machines"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "PyQt6==6.6.1",
    "paramiko==3.4.0",
    "gitpython==3.1.41"
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
pip = ">=23.3.2" 